chore: fix bug(s); format file(s) and code
