+++ src/database/mod.rs
@@ -1,4 +1,2 @@
 pub mod display;
-pub mod edit;
 pub mod format_md;
-pub mod get_notes;
