+++ src/note.rs
@@ -1,8 +1,9 @@
 use crate::{
-    cursor_to_origin,
-    database::{display::display, get_notes::get_notes},
+    api, cursor_to_origin,
+    database::display::display,
+    multiselect,
     prompts::{confirm::confirm, input::input, select::select},
-    truncated_note,
+    truncate_note,
 };
 use async_std::path::PathBuf;
 use chrono::prelude::Local;
@@ -18,8 +19,9 @@ pub struct Note {
 
 impl Note {
     pub fn create(db_file: &PathBuf) -> Result<(), Box<dyn std::error::Error>> {
+        cursor_to_origin()?;
         let mut inputted_note = Note {
-            id: get_notes(db_file)?.len(),
+            id: api::get_notes(db_file)?.len(),
             name: input("Name:", "".to_string()),
             content: input("Content:", "".to_string()),
             created: format!("{}", Local::now().format("%A %e %B, %H:%M")),
@@ -52,8 +54,8 @@ impl Note {
         }
     }
 
-    fn show(db_file: &PathBuf) -> Result<(), Box<dyn std::error::Error>> {
-        let saved_notes = get_notes(db_file)?;
+    pub fn show(db_file: &PathBuf) -> Result<(), Box<dyn std::error::Error>> {
+        let saved_notes = api::get_notes(db_file)?;
 
         if saved_notes.is_empty() {
             println!("You don't have any notes.");
@@ -61,18 +63,19 @@ impl Note {
         }
 
         let mut options: Vec<String> = Vec::new();
-        truncated_note(&mut options, db_file)?;
+        truncate_note(&mut options, db_file)?;
         let selection = select("Select the note that you want to view:", &options);
         let mut selected_note = &saved_notes[selection];
+        cursor_to_origin()?;
 
         display(&mut selected_note)?;
         Ok(())
     }
 
     pub fn edit(db_file: &PathBuf) -> Result<(), Box<dyn std::error::Error>> {
-        let saved_notes = get_notes(db_file)?;
+        let saved_notes = api::get_notes(db_file)?;
         let mut options: Vec<String> = Vec::new();
-        truncated_note(&mut options, db_file)?;
+        truncate_note(&mut options, db_file)?;
         let selection = select("Select the note that you want to edit:", &options);
         let selected_note = &saved_notes[selection];
         let updated_note = Note {
@@ -97,10 +100,43 @@ impl Note {
                 )?;
 
                 cursor_to_origin()?;
+                api::save_note(&updated_note, db_file)?; // why the fuck whas this line not here yet
                 println!("Note updated successfully.");
                 Ok(())
             }
             false => Ok(()),
+        }
+    }
+
+    pub fn delete(db_file: &PathBuf) -> Result<(), Box<dyn std::error::Error>> {
+        // yep yep yeah
+        if api::get_notes(db_file)?.is_empty() {
+            println!("You can't delete notes, because there are none.");
+            return Ok(());
+        }
+
+        let mut options: Vec<String> = Vec::new();
+        truncate_note(&mut options, db_file)?;
+        let selections = multiselect(
+            "Select the note(s) that you want to delete:\nSpace to select, Enter to confirm.",
+            options,
+        );
+
+        let mut prompt = "Are you sure that you want to delete these notes?";
+        if selections.len() == 1 {
+            prompt = "Are you sure that you want to delete this note?";
+        }
+
+        cursor_to_origin()?;
+        if selections.is_empty() {
+            println!("You didn't select any notes.");
+            return Ok(());
+        } else {
+            if confirm(prompt) {
+                api::delete_notes(selections, db_file)?;
+            }
+            println!("Notes deleted successfully.");
+            return Ok(());
         }
     }
 }
